+++ plugin/publish/delete.ts
