+++ plugin/publish/filesManagement.ts
